backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.10
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.0.0
requests==2.28.2
httpx==0.24.0
pytest==7.3.1
cython==0.29.32
alembic==1.10.3
python-dotenv==1.0.0
